# public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "SABERHUB Core LMS API",
    "description": "API REST robusta y estandarizada de SABERHUB para la integración de servicios educativos, automatización de inscripciones, reporte de progreso (SCORM), videoconferencias en vivo (Google Meet / Zoom) y webhooks transaccionales en tiempo real.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/api",
      "description": "Servidor Local"
    }
  ],
  "paths": {
...
    "/cursos/{id}/sesiones": {
      "post": {
        "summary": "Programar Videoconferencia",
        "description": "Programar una nueva clase sincrónica por Zoom o Google Meet en un curso específico, notificando por SMTP a los estudiantes inscritos.",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["titulo", "fechaInicio"],
                "properties": {
                  "titulo": { "type": "string", "example": "Introducción al Módulo Sincrónico" },
                  "descripcion": {
                    "type": "string",
                    "example": "Resolveremos dudas sobre el estándar SCORM"
                  },
                  "fechaInicio": {
                    "type": "string",
                    "format": "date-time",
                    "example": "2026-06-01T15:00:00Z"
                  },
                  "duracion": {
                    "type": "integer",
                    "example": 60,
                    "description": "Duración estimada en minutos"
                  },
                  "plataforma": {
                    "type": "string",
                    "enum": ["meet", "zoom"],
                    "default": "meet",
                    "example": "zoom"
                  },
                  "urlReunion": {
                    "type": "string",
                    "description": "URL personalizada (opcional, si no se envía se autogenera una de Zoom/Meet)",
                    "example": ""
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Videoconferencia programada y notificada por correo exitosamente."
          },
          "401": { "description": "No autorizado" },
          "403": { "description": "Permisos insuficientes" }
        }
      }
    },
...
    "/progreso/scorm": {
      "get": {
        "summary": "Obtener Progreso SCORM",
        "description": "Obtener las variables CMI de seguimiento guardadas de un estudiante para un SCORM específico.",
        "parameters": [
          { "name": "leccionId", "in": "query", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "Progreso SCORM retornado correctamente.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "cmiData": { "type": "object" }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Guardar Progreso SCORM",
        "description": "Sincroniza y guarda variables CMI (SCORM 1.2 o 2004). Si detecta estado completado/aprobado, completa automáticamente la lección del LMS y evalúa si otorga certificado. Si se otorga, dispara el webhook `certificacion.emitida`.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["leccionId", "cmiData"],
                "properties": {
                  "leccionId": { "type": "string", "example": "leccion-cuid-789" },
                  "cmiData": {
                    "type": "object",
                    "properties": {
                      "cmi.core.lesson_status": {
                        "type": "string",
                        "enum": ["completed", "passed", "incomplete", "failed"]
                      },
                      "cmi.core.lesson_location": { "type": "string" },
                      "cmi.completion_status": { "type": "string" }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Progreso guardado y sincronizado con el LMS." }
        }
      }
    },
...
  },
...
}